requirements/common.txt
### COMMON REQUIREMENTS ###
# Django base
django>=1.5,<1.6
ipython>=0.13.1,<0.14
South==0.8.1

# Django utils
django-crispy-forms
django_select2
django-classy-tags>=0.3.1
django-sekizai
fancy_tag
django-braces
django-model-utils
django-dbbackup

# Other libs probably needed
fabtools
sh
pillow>=2.0.0
unicode-slugify
python-dateutil

# We don't want a different db even in development
psycopg2==2.5

# llazzaro / django-scheduler , calendaring app for Django
pytz>=2013b
-e git+https://github.com/llazzaro/django-scheduler.git#egg=django-scheduler

django-inplaceedit>=1.2,<1.3
